Adiciona tab bar, nav bar items e novas cervejas

+++ ArrotoAmargo/Modelos/AvaliacaoCerveja.swift
@@ -16,7 +16,7 @@ struct AvaliacaoCerveja: Hashable, Codable, Identifiable {
     var nomeFamiliaCerveja: String
     var nomeCervejaria: String
     var nota: Int
-    var ibu: Int
+    var ibu: Float
     var teorAlcoolico: Float
     
     var locationCoordinate: CLLocationCoordinate2D {
